requirements.txt
# Requirements exported 2025-11-19T04:29:09.534477
backports.tarfile==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
importlib_metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
keyring==25.6.0
more-itertools==10.8.0
pillow==12.0.0
pycparser==2.23
pyenchant==3.3.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pysqlcipher3==1.2.0
requests==2.32.5
SecretStorage==3.4.0
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
